backend/pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = "Backend for AWSome Generative Drinks App"
authors = ["Ana Cunha, Marcelo Palladino"]

[tool.poetry.dependencies]
python = "^3.12"
aws-lambda-powertools = {version = "^2.43.1", extras = ["tracer"]}
pydantic = "^2.9.2"
boto3 = "^1.37.19"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
aws-cdk-lib = "^2.160.0"
constructs = "^10.3.0"
"aws-cdk-aws-lambda-python-alpha" = "^2.160.0a0"
gitpython = "^3.1.43"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
